Backend/requirements.txt
faiss_cpu==1.10.0
Flask==3.1.2
flask_cors==6.0.2
groq==1.0.0
numpy==2.4.2
sentence_transformers==4.1.0
